Issue #9 - added basic blocks for Sensor detail and create page. Need to implement other ones before though.

+++ SmartHome.DomainCore/Data/EntityReference.cs
@@ -2,10 +2,10 @@ namespace SmartHome.DomainCore.Data
 {
     public class EntityReference
     {
-        public int Id { get; }
+        public long Id { get; }
         public string? Name { get;  }
         
-        public EntityReference(int id, string name)
+        public EntityReference(long id, string name)
         {
             Id = id;
             Name = name;
